fixed database

+++ api-backend/utilities/database.js
@@ -11,10 +11,10 @@ const dbconnection = {
 }
 
 const dbInit = () => {
+    let db;
     
     // Connect to MySQL server
-    const connection = mysql.createConnection(
-        {
+    const connection = mysql.createConnection({
             host: dbconnection.host,
             user: dbconnection.user,
             password: dbconnection.password,
@@ -24,15 +24,17 @@ const dbInit = () => {
     
     // Create database if it does not exist
     connection.query('CREATE DATABASE IF NOT EXISTS softeng2234', (err, results) => {
-        if(err) console.error(err);
-        else console.log(results);
-    });
-
-    // Connect sequelize to database
-    const db = new Sequelize('softeng2234', dbconnection.user, dbconnection.password, {
-        host: dbconnection.host,
-        port: dbconnection.port,
-        dialect: dbconnection.dialect
+        if(err) {
+            db = undefined;
+            console.error(err);
+        } else {
+            // Connect sequelize to database
+            db = new Sequelize('softeng2234', dbconnection.user, dbconnection.password, {
+                host: dbconnection.host,
+                port: dbconnection.port,
+                dialect: dbconnection.dialect
+            });
+        }
     });
 
     return db;
@@ -42,6 +44,7 @@ const db = dbInit();
 
 verifyDB = async () => {
     try {
+        if(!db) throw err;
         await db.authenticate();
         await db.sync({force: false, alter: false});
         await Administrator.findOrCreate({where: {username: 'admin'}, defaults: {
@@ -67,13 +70,15 @@ verifyDB = async () => {
             education: 'Upper Secondary Education',
             income: '<5.000'
         }});
-        await User.create({username: 'user1', password: 'user1'});
-        await User.create({username: 'user2', password: 'user2'});
-        await User.create({username: 'user3', password: 'user3'});
-        await User.create({username: 'user4', password: 'user4'});
-        await User.create({username: 'user5', password: 'user5'});
+        // await User.create({username: 'user1', password: 'user1'});
+        // await User.create({username: 'user2', password: 'user2'});
+        // await User.create({username: 'user3', password: 'user3'});
+        // await User.create({username: 'user4', password: 'user4'});
+        // await User.create({username: 'user5', password: 'user5'});
         console.log('Connection with the database has been established successfully');
-    } catch(err) {console.error('Unable to connect to the database:', err);}
+    } catch(err) {
+        console.error('Unable to connect to the database:', err);
+    }
 }
 
 // Models
